Expose more problems plus minor cleanup around.

+++ src/algorithm/sa_corana.cpp
@@ -126,8 +126,8 @@ void sa_corana::evolve(population &pop) const {
 
 	//Main SA loops
 	for (size_t jter = 0; jter < n_T; ++jter) {
-		for (size_t mter = 0; mter < m_step_adj; ++mter) {
-			for (size_t kter = 0; kter < m_bin_size; ++kter) {
+		for (int mter = 0; mter < m_step_adj; ++mter) {
+			for (int kter = 0; kter < m_bin_size; ++kter) {
 				size_t nter =  (size_t)(m_drng() * Dc);
 				for (size_t numb = 0; numb < Dc ; ++numb) {
 					nter = (nter + 1) % Dc;
